Fix 'possible loss of data' warning
Qt doesn't like to use size_t anywhere.

+++ obs/window-basic-main.cpp
@@ -1776,7 +1776,7 @@ void OBSBasic::UploadLog(const char *file)
 		return;
 	}
 
-	logUploadPostData.setData(json, strlen(json));
+	logUploadPostData.setData(json, (int)strlen(json));
 
 	QUrl url("https://api.github.com/gists");
 	logUploadReply = networkManager.post(QNetworkRequest(url),
